+++ R/plotting-helpers.R
@@ -1,9 +1,8 @@
 ### deal with spurious data.table warnings
 if (getRversion() >= "3.1.0") {
-  utils::globalVariables(c(".", ".N"))
+  utils::globalVariables(c(".", ".N", "keepAll"))
 }
 
-################################################################################
 #' Find the number of layers in a Spatial Object
 #'
 #' There are already methods for \code{Raster*} in the raster package.
@@ -88,7 +87,6 @@ setMethod(
     return(1L)
 })
 
-################################################################################
 #' Extract the layer names of Spatial Objects
 #'
 #' There are already methods for \code{Raster*} objects. This adds methods for
@@ -191,7 +189,6 @@ setMethod(
     return("")
 })
 
-################################################################################
 #' Assess whether a list of extents are all equal
 #'
 #' @param extents list of extents objects
@@ -233,7 +230,6 @@ setMethod(
     )
 })
 
-################################################################################
 #' Make a \code{.quickPlot} class object
 #'
 #' Builds a \code{.quickPlot} object from a list of objects.
@@ -353,6 +349,7 @@ setMethod(
       } else {
         quickPlotGrobList[[lN[x]]] <- new(".quickPlotGrob")
         quickPlotGrobList[[lN[x]]]@plotArgs <- lapply(plotArgs, function(y) y[[x]])
+        quickPlotGrobList[[lN[x]]]@plotArgs$zoomExtent <- plotArgs$zoomExtent[[x]]
         quickPlotGrobList[[lN[x]]]@plotArgs$gpText <- plotArgs$gpText[x]
         quickPlotGrobList[[lN[x]]]@plotArgs$gpAxis <- plotArgs$gpAxis[x]
         quickPlotGrobList[[lN[x]]]@plotArgs$gp <- plotArgs$gp[x]
@@ -396,7 +393,6 @@ setMethod(
     return(newPlots)
 })
 
-################################################################################
 #' Convert \code{plotArgs} to list of lists
 #'
 #' Internal function. Take the inputs as plotArgs to the Plot function, and make
@@ -638,7 +634,6 @@ setMethod(
     return(p)
 })
 
-################################################################################
 #' Make \code{SpatialLines} object from two \code{SpatialPoints} objects
 #'
 #' The primary conceived usage of this is to draw arrows following the
@@ -678,7 +673,6 @@ setMethod(
     }), proj4string = crs(from)) # nolint
 })
 
-################################################################################
 #' Parse arguments and find environments
 #'
 #' Internal function used within objectNames.
@@ -852,7 +846,6 @@ setMethod(
               envs = envs))
 }
 
-################################################################################
 #' Parsing of elements
 #'
 #' This is a generic definition that can be extended according to class. Intended
@@ -948,7 +941,6 @@ setMethod("gpar",
             return(grid::gpar(...))
 })
 
-################################################################################
 #' Internal functions used by Plot
 #'
 #' Extract colors, legends and other things from object, and convert to a plotGrob
@@ -998,13 +990,46 @@ setMethod(
     return(zMat)
 })
 
+#' @importFrom rgeos gIntersects gArea
 setMethod(
   ".preparePlotGrob",
   signature = c("Spatial", ".quickPlotGrob"),
   definition = function(grobToPlot, sGrob, takeFromPlotObj, arr, newArr,
                         quickPlotGrobCounter, subPlots, cols) {
-    if (!is.null(sGrob@plotArgs$zoomExtent)) {
-      grobToPlot <- crop(grobToPlot, sGrob@plotArgs$zoomExtent)
+
+    if (!is.null(sGrob@plotArgs$zoomExtent) &&
+        !identical(extent(grobToPlot), arr@extents[[subPlots]])) {
+        #!identical(arr@extents[[subPlots]], sGrob@plotArgs$zoomExtent)) {
+      useCrop <- FALSE
+      if (!useCrop) {
+        zoom <- sGrob@plotArgs$zoomExtent
+        extPolygon <- as(zoom, "SpatialPolygons")
+        crs(extPolygon) <- crs(grobToPlot)
+        extPolygon <- list(extPolygon)
+        names(extPolygon) <- sGrob@plotName
+
+        fullArea <-
+          rgeos::gArea(as(extent(grobToPlot), "SpatialPolygons"))
+        zoomArea <-
+          rgeos::gArea(as(extent(zoom), "SpatialPolygons"))
+        numPolys <- length(grobToPlot)
+        ratio <- fullArea / zoomArea
+        if (numPolys / ratio * 5 > getOption("quickPlot.maxNumPolygons", 3e3)) {
+          polySeq <-
+            .polygonSeq(grobToPlot,
+                        maxNumPolygons = getOption("quickPlot.maxNumPolygons", 3e3))
+          .showingOnlyMessage(numShowing = getOption("quickPlot.maxNumPolygons", 3e3),
+                              totalAvailable = length(grobToPlot))
+          grobToPlot <- grobToPlot[polySeq,]
+
+        }
+        message("Cropping to new extent")
+        a <- rgeos::gIntersects(grobToPlot, extPolygon[[1]], byid = TRUE)
+        grobToPlot <- grobToPlot[a[1,],]
+      } else {
+        grobToPlot <- crop(grobToPlot, sGrob@plotArgs$zoomExtent)
+      }
+
     }
 
     # This handles SpatialPointsDataFrames with column "color"
@@ -1014,8 +1039,7 @@ setMethod(
     zMat <- list(z = grobToPlot, minz = 0, maxz = 0,
                  cols = sGrob@plotArgs$cols, real = FALSE)
     return(zMat)
-  })
-
+})
 
 setMethod(
   ".preparePlotGrob",
@@ -1027,7 +1051,7 @@ setMethod(
 
     list(z = grobToPlot, minz = 0, maxz = 0,
          cols = sGrob@plotArgs$cols, real = FALSE)
-  })
+})
 
 setMethod(
   ".preparePlotGrob",
@@ -1042,8 +1066,6 @@ setMethod(
   return(zMat)
 })
 
-
-################################################################################
 #' @param whPlotFrame Numeric. Which plot within the quickPlotGrobPlots object.
 #'
 #' @include plotting-classes.R
@@ -1379,7 +1401,6 @@ setMethod(
     return(sGrob)
 })
 
-
 #' @include plotting-classes.R
 #' @aliases PlotHelpers
 #' @keywords internal
@@ -1423,7 +1444,6 @@ setMethod(
     return(sGrob)
 })
 
-################################################################################
 #' Identify where to get the grob from
 #'
 #' Internal function.
@@ -1481,7 +1501,6 @@ setMethod(
     return(grobToPlot)
 })
 
-################################################################################
 #' Prepare raster for plotting
 #'
 #' Internal function. Takes a raster .quickPlotGrob, and converts zoomExtent into
@@ -1539,8 +1558,6 @@ setMethod(
               legendRange = legendRange, zoom = zoom))
 }
 
-
-################################################################################
 #' Merge two quickPlot objects
 #'
 #' Merges two \code{.quickPlot} objects
@@ -1690,9 +1707,8 @@ setMethod(
         lapply(x, function(y) TRUE)
       })
     ))
-  })
+})
 
-################################################################################
 #' Determine optimal plotting arrangement of plot objects
 #'
 #' Internal function. Assesses the device geometry, the map geometry, and the
@@ -1776,8 +1792,6 @@ setMethod(
     return(out)
 })
 
-
-################################################################################
 #' Plot spatial grobs (using \pkg{grid} package)
 #'
 #' Internal function. Plot a raster Grob, a points Grob, polygon Grob.
@@ -1795,7 +1809,7 @@ setMethod(
 #' plotting down.
 #'
 #' The suggested package \code{fastshp} can be installed with:
-#' \code{install.packages("fastshp", repos = "http://rforge.net", type = "source")}.
+#' \code{install.packages("fastshp", repos = "https://rforge.net", type = "source")}.
 #'
 #' NOTE: you may get errors relating to not having installed the software tools
 #' required for building R packages on your system.
@@ -1848,6 +1862,8 @@ setMethod(
 #' @param ...     Additional arguments. None currently implemented.
 #'
 #' @author Eliot McIntire
+#' @export
+#' @exportMethod .plotGrob
 #' @importFrom data.table ':=' data.table
 #' @importFrom grDevices as.raster
 #' @importFrom grid gpar gTree gList rasterGrob textGrob grid.draw
@@ -1855,8 +1871,6 @@ setMethod(
 #' @importFrom raster extent pointDistance xmin xmax ymin ymax
 #' @keywords internal
 #' @rdname plotGrob
-#' @export
-#' @exportMethod .plotGrob
 #'
 setGeneric(
   ".plotGrob",
@@ -1866,7 +1880,6 @@ setGeneric(
   standardGeneric(".plotGrob")
 })
 
-############## SpatialPoints - thin
 #' @rdname plotGrob
 #' @importFrom grid pointsGrob
 setMethod(
@@ -1909,7 +1922,7 @@ setMethod(
           message(
             paste(
               "To speed up Polygons plotting using Plot install the fastshp package:\n",
-              "install.packages(\"fastshp\", repos=\"http://rforge.net\", type=\"source\")."
+              "install.packages(\"fastshp\", repos=\"https://rforge.net\", type=\"source\")."
             )
           )
           if (Sys.info()[["sysname"]] == "Windows") {
@@ -2049,7 +2062,6 @@ setMethod(
     return(invisible())
 })
 
-################################################################################
 #' @rdname plotGrob
 #' @importFrom grid polygonGrob
 setMethod(
@@ -2069,20 +2081,15 @@ setMethod(
           xmax(extent(grobToPlot)) - xmin(extent(grobToPlot))) / 2.4e4
     }
 
+    if (is.null(gp$fill)) {
+      gp$fill <- rep(RColorBrewer::brewer.pal(8, "Set2"), length.out = length(grobToPlot))
+    }
+
     # For speed of plotting
     xyOrd <- quickPlot::thin(grobToPlot, tolerance = speedupScale * speedup,
-                             returnDataFrame = TRUE, minCoordsToThin = 1e3)
+                             returnDataFrame = TRUE, minCoordsToThin = 1e5, ...)
 
-    gp$fill[xyOrd[["hole"]]] <- "#FFFFFF00"
-    polyGrob <- gTree(children = gList(
-      polygonGrob(
-        x = xyOrd[["xyOrd"]]$x, y = xyOrd[["xyOrd"]]$y,
-        id.lengths = xyOrd[["idLength"]]$V1,
-        gp = gp, default.units = "native"
-      )
-    ),
-    gp = gp,
-    cl = "plotPoly")
+    polyGrob <- .createPolygonGrob(gp = gp, xyOrd = xyOrd)
     grid.draw(polyGrob, recording = FALSE)
     return(invisible(polyGrob))
 })
@@ -2134,7 +2141,7 @@ setMethod(
           message(
             paste(
               "To speed up Lines plotting using Plot, install the fastshp package:\n",
-              "install.packages(\"fastshp\", repos=\"http://rforge.net\", type=\"source\")"
+              "install.packages(\"fastshp\", repos=\"https://rforge.net\", type=\"source\")"
             )
           )
           if (Sys.info()[["sysname"]] == "Windows") {
@@ -2174,7 +2181,6 @@ setMethod(
     return(invisible(lineGrob))
 })
 
-################################################################################
 #' Make an optimal layout of plots
 #'
 #' Internal function. Using the size of the current device, and number and
@@ -2195,11 +2201,11 @@ setMethod(
 #'              written above plots and should be included as part of layout
 #'               calculation. Default is \code{TRUE}.
 #'
-#' @include plotting-classes.R
+#' @author Eliot McIntire
 #' @importFrom grid unit unit.c
+#' @include plotting-classes.R
 #' @keywords internal
 #' @rdname makeLayout
-#' @author Eliot McIntire
 #'
 .makeLayout <- function(arr, visualSqueeze,
                         legend = TRUE, axes = TRUE, title = TRUE) {
@@ -2240,7 +2246,6 @@ setMethod(
               visualSqueeze = visualSqueeze))
 }
 
-################################################################################
 #' Make viewports
 #'
 #' Given a set of extents, and a layout for these extents, this function will
@@ -2263,7 +2268,6 @@ setMethod(
 #' @rdname makeViewports
 #'
 .makeViewports <- function(sPlot, newArr = FALSE) {
-
   arr <- sPlot@arr
   sgl <- sPlot@quickPlotGrobList
 
@@ -2286,7 +2290,6 @@ setMethod(
           # for non spatial objects
           extent(c(xmin = 0, xmax = 2, ymin = 0, ymax = 2))
         }
-
       }
     }))
   }))
@@ -2442,33 +2445,38 @@ sp2sl <- function(sp1, from) {
   SpatialLines(l)
 }
 
-
-
-#' Thin a polygon using fastshp::thin
+#' Thin a polygon using \code{fastshp::thin}
 #'
 #' For visualizing, it is sometimes useful to remove points in Spatial* objects.
-#' This will change the geometry, so it is not recommended for computation. This
-#' is similar to rgeos::gSimplify and sf::st_simplify, but faster than both (see examples)
-#' for large shapefiles, particularly if \code{returnDataFrame} is \code{TRUE}.
-#' \code{thin} will not attempt to preserve topology.
+#' This will change the geometry, so it is not recommended for computation.
+#' This is similar to \code{rgeos::gSimplify} and \code{sf::st_simplify},
+#' but faster than both (see examples) for large shapefiles, particularly if
+#' \code{returnDataFrame} is \code{TRUE}.
+#' \emph{\code{thin} will not attempt to preserve topology.}
 #' It is strictly for making smaller polygons for the purpose (likely)
 #' of visualizing more quickly.
 #'
 #' @param x A Spatial* object
-#' @param returnDataFrame If \code{TRUE}, this will return a list of 3 elements, xyOrd, hole,
-#'        idLength. If \code{FALSE}, the default, it will return a \code{SpatialPolygons}
-#'        object
+#' @param tolerance Maximum allowable distance for a point to be removed.
+#' @param returnDataFrame If \code{TRUE}, this will return a list of 3 elements,
+#'        \code{xyOrd}, \code{hole}, and \code{idLength}.
+#'        If \code{FALSE} (default), it will return a \code{SpatialPolygons} object.
 #' @param minCoordsToThin If the number of coordinates is smaller than this number,
 #'        then thin will just pass through, though it will take the time required to
 #'        calculate how many points there are (which is not NROW(coordinates(x)) for
 #'        a SpatialPolygon)
+#' @param ... Passed to methods (e.g., \code{maxNumPolygons})
+#' @param maxNumPolygons For speed, \code{thin} can also simply remove some of the
+#'        polygons. This is likely only a reasonable thing to do if there are
+#'        a lot of polygons being plotted in a small space. Current default is
+#'        taken from \code{options('quickPlot.maxNumPolygons')}, with a message.
 #'
-#' @inheritParams fastshp::thin
-#' @rdname thin
-#' @importFrom data.table set data.table as.data.table
-#' @importFrom sp Polygon Polygons SpatialPolygons CRS SpatialPolygonsDataFrame
-#' @importFrom raster xmax xmin
 #' @export
+#' @importFrom data.table as.data.table data.table set
+#' @importFrom raster xmax xmin
+#' @importFrom sp CRS Polygon Polygons SpatialPolygons SpatialPolygonsDataFrame
+#' @rdname thin
+#'
 #' @examples
 #' library(raster)
 #'
@@ -2512,15 +2520,17 @@ sp2sl <- function(sp1, from) {
 #' clearPlot()
 #' Plot(a)
 #' NROW(a@polygons[[1]]@Polygons[[1]]@coords)
-#' aThin <- thin(a, 200)
-#' NROW(aThin@polygons[[1]]@Polygons[[1]]@coords) # fewer
-#' Plot(aThin) # looks similar
+#' if (require(fastshp)) {
+#'   aThin <- quickPlot::thin(a, 200)
+#'   NROW(aThin@polygons[[1]]@Polygons[[1]]@coords) # fewer
+#'   Plot(aThin) # looks similar
+#' }
 #'
 #' # compare -- if you have rgeos
 #' # if (require("rgeos")) {
 #' #   aSimplify <- gSimplify(a, tol = 200)
 #' #   NROW(aSimplify@polygons[[1]]@Polygons[[1]]@coords) # fewer
-#' #  Plot(aSimplify)
+#' #   Plot(aSimplify)
 #' # }
 #'
 #' # compare -- if you have sf
@@ -2533,7 +2543,8 @@ sp2sl <- function(sp1, from) {
 #' # }
 #'
 #' # thin is faster than rgeos::gSimplify and sf::st_simplify on large shapefiles
-#' \dontrun{ # this involves downloading a 9 MB file
+#' \dontrun{
+#'   # this involves downloading a 9 MB file
 #'   setwd(tempdir())
 #'   albertaEcozoneFiles <- c("Natural_Regions_Subregions_of_Alberta.dbf",
 #'                            "Natural_Regions_Subregions_of_Alberta.lyr",
@@ -2569,23 +2580,25 @@ sp2sl <- function(sp1, from) {
 #'    # st_simplify(aSF, dTolerance = 10) 4087.343 4344.936 4910.299     6   b
 #'   #}
 #' }
-thin <- function(x, tolerance, returnDataFrame, minCoordsToThin) {
+thin <- function(x, tolerance, returnDataFrame, minCoordsToThin, ...) {
   UseMethod("thin")
 }
 
-#' @rdname thin
 #' @export
-thin.SpatialPolygons <- function(x, tolerance = NULL, returnDataFrame = FALSE, minCoordsToThin = 0) {
+#' @rdname thin
+thin.SpatialPolygons <- function(x, tolerance = NULL, returnDataFrame = FALSE, minCoordsToThin = 0,
+                                 maxNumPolygons = getOption("quickPlot.maxNumPolygons", 3e3), ...) {
 
   # For speed of plotting
   xyOrd <- .fortify(x, matchFortify = FALSE,
-                    simple = returnDataFrame) # a list: out, hole, idLength
+                    simple = returnDataFrame, maxNumPolygons) # a list: out, hole, idLength
   if (is.null(tolerance)) {
     tolerance <- (raster::xmax(x) - raster::xmin(x)) * 0.0001
     message("tolerance set to ", tolerance)
   }
-  if (requireNamespace("fastshp")) {
+  if (requireNamespace("fastshp", quietly = TRUE)) {
     if (NROW(xyOrd[["out"]]) > minCoordsToThin) {
+      message("Some polygons have been simplified")
       thinRes <- fastshp::thin(xyOrd[["out"]]$x, xyOrd[["out"]]$y,
                              tolerance = tolerance, id = xyOrd[["out"]]$groups)
 
@@ -2596,12 +2609,11 @@ thin.SpatialPolygons <- function(x, tolerance = NULL, returnDataFrame = FALSE, m
 
       #xyOrd[["out"]] <- xyOrd[["out"]][thinRes, ]# thin line
       if (returnDataFrame) {
-        xyOrd[["idLength"]] <- xyOrd[["out"]][,list(V1=.N),by=groups]
+        xyOrd[["idLength"]] <- xyOrd[["out"]][, list(V1 = .N), by = groups]
       } else {
         # clean up a bit
         set(xyOrd[["out"]], , "order", NULL)
         set(xyOrd[["out"]], , "groups", NULL)
-
 
         polyList <- split(xyOrd[["out"]], by = c("Polygons", "Polygon"),
                            flatten = FALSE, keep.by = FALSE)
@@ -2614,19 +2626,26 @@ thin.SpatialPolygons <- function(x, tolerance = NULL, returnDataFrame = FALSE, m
           Polygons(poly, ID = outerI)
         })
 
+        names1 <- unique(xyOrd$out$Polygons)
         xyOrd <- SpatialPolygons(bb, proj4string = CRS(proj4string(x)))
-        if (is(x, "SpatialPolygonsDataFrame"))
-          xyOrd <- SpatialPolygonsDataFrame(xyOrd, data = x@data)
-        return(xyOrd)
+        if (is(x, "SpatialPolygonsDataFrame")) {
+          if (length(x) > maxNumPolygons) {
+            dat <- x@data[as.numeric(names1) + 1,]
+            #row.names(dat) <- as.character(seq_len(length(xyOrd)))
+          } else {
+            dat <- x@data
+          }
+          xyOrd <- SpatialPolygonsDataFrame(xyOrd, data = dat)
+        }
 
+        return(xyOrd)
       }
     }
   } else {
-
     message(
       paste(
         "To speed up Polygons plotting using Plot install the fastshp package:\n",
-        "install.packages(\"fastshp\", repos=\"http://rforge.net\", type=\"source\")."
+        "install.packages(\"fastshp\", repos=\"https://rforge.net\", type=\"source\")."
       )
     )
     if (Sys.info()[["sysname"]] == "Windows") {
@@ -2642,10 +2661,9 @@ thin.SpatialPolygons <- function(x, tolerance = NULL, returnDataFrame = FALSE, m
                 idLength = xyOrd[["idLength"]])
 }
 
-
-#' @rdname thin
 #' @export
-thin.default <- function(x, tolerance, returnDataFrame, minCoordsToThin) {
+#' @rdname thin
+thin.default <- function(x, tolerance, returnDataFrame, minCoordsToThin, ...) {
   message("No method for that class of object exists. See methods('thin') to see current methods")
 }
 
@@ -2657,41 +2675,47 @@ thin.default <- function(x, tolerance, returnDataFrame, minCoordsToThin) {
 #' @name fortify
 #' @importFrom data.table setDT set
 #' @keywords internal
-.fortify <- function(x, matchFortify = TRUE, simple = FALSE) {
+.fortify <- function(x, matchFortify = TRUE, simple = FALSE,
+                     maxNumPolygons = getOption("quickPlot.maxNumPolygons", 3e3)) {
   ord <- x@plotOrder
+  if (length(ord) > maxNumPolygons) {
+
+    polygonSeq <- .polygonSeq(x, maxNumPolygons) #if (is.numeric(x@data$Shape_Area)) {
+    ord <- ord[polygonSeq]
+    .showingOnlyMessage(numShowing = maxNumPolygons,
+                        totalAvailable = length(x@plotOrder))
+  }
   ordSeq <- seq(ord)
+
   xy <- lapply(ordSeq, function(i) {
-    lapply(x@polygons[[i]]@Polygons, function(j) {
+    lapply(x@polygons[[ord[i]]]@Polygons, function(j) {
       j@coords
     })
   })
 
   hole <- tryCatch(unlist(lapply(ordSeq, function(xx) {
-    lapply(x@polygons[[xx]]@Polygons, function(yy)
+    lapply(x@polygons[[ord[xx]]]@Polygons, function(yy)
       yy@hole)
   })), error = function(xx) FALSE)
 
   IDs <- tryCatch(unlist(lapply(ordSeq, function(xx) {
-    x@polygons[[xx]]@ID
+    x@polygons[[ord[xx]]]@ID
   })), error = function(xx) FALSE)
 
-
   ordInner <- lapply(ordSeq, function(xx) {
-    x@polygons[[xx]]@plotOrder
+    x@polygons[[ord[xx]]]@plotOrder
   })
 
   xyOrd.l <- lapply(ordSeq, function(i) { # nolint
-    xy[[i]][ordInner[[i]]]
+    xy[[ordSeq[i]]][ordInner[[ordSeq[i]]]]
   })
 
   idLength <- data.table(V1 = unlist(lapply(xyOrd.l, function(i) {
     lapply(i, length)
   })) / 2)
 
-
   numPolygons <- unlist(length(xyOrd.l))
   numPolygon <- unlist(lapply(xyOrd.l, length))
-
 
   xyOrd <- do.call(rbind, lapply(xyOrd.l, function(i) {
     do.call(rbind, i)
@@ -2724,4 +2748,33 @@ thin.default <- function(x, tolerance, returnDataFrame, minCoordsToThin) {
 
     return(out)
   }
+}
+
+.polygonSeq <- function(polygon, maxNumPolygons) {
+  if (is.numeric(polygon@data$Shape_Area)) {
+    which(polygon@data$Shape_Area>(sort(polygon@data$Shape_Area, decreasing = TRUE)[maxNumPolygons]))
+  } else {
+    round(seq(1, length(polygon), length.out = maxNumPolygons))
+  }
+
+}
+
+.createPolygonGrob <- function(gp, xyOrd) {
+  gp$fill[xyOrd[["hole"]]] <- "#FFFFFF00"
+  polyGrob <- gTree(children = gList(
+    polygonGrob(
+      x = xyOrd[["xyOrd"]]$x, y = xyOrd[["xyOrd"]]$y,
+      id.lengths = xyOrd[["idLength"]]$V1,
+      gp = gp, default.units = "native"
+    )
+  ),
+  gp = gp,
+  cl = "plotPoly")
+
+}
+
+.showingOnlyMessage <- function(numShowing, totalAvailable) {
+  message("Showing only ", numShowing, " of ",
+          totalAvailable," polygons in this view. See options('quickPlot.maxNumPolygons')")
+
 }