+++ NEWS.md
@@ -1,7 +1,18 @@
 Known issues: https://github.com/PredictiveEcology/quickPlot/issues
 
+version 0.1.4
+=============
+* much faster for `SpatialPolygons`
+* `clickExtent` now more interactive with new examples
+* remove dependency on `reproducible`
+* `SpatialPolygons` get a fill colour by default if not passed in by user
+* fix bug related to plotting with NAs (#14, @CeresBarros)
+* bug fixes
+* new options("quickPlot.maxNumPolygons"), used for plotting of large `SpatialPolygons` objects
+ 
+
 version 0.1.3
-==================
+=============
 * new function, `thin`, which is similar to rgeos::gSimplify, but uses fastshp::thin, and can get much faster speeds, 
   used internally for plotting, but also available as a function to user
 * minor bug fix for when quickPlot::Plot is called inside a custom function called "Plot"